refactor: switch to node:sqlite
