+++ package.json
@@ -34,9 +34,9 @@
     "@atcute/client": "^4.2.1",
     "@atcute/identity-resolver": "1.2.2",
     "@atcute/identity-resolver-node": "1.0.3",
+    "@atcute/jetstream": "^1.1.2",
     "@atcute/lexicons": "1.2.9",
     "@atcute/oauth-node-client": "^1.1.0",
-    "@atcute/jetstream": "^1.1.2",
     "@atcute/tid": "1.1.2",
     "@hono/node-server": "^1.19.7",
     "@hono/prometheus": "^1.0.2",
@@ -48,10 +48,8 @@
     "@opentelemetry/sdk-trace-node": "1.29.0",
     "@opentelemetry/semantic-conventions": "1.28.0",
     "@tanstack/react-query": "^5.90.12",
-    "@types/better-sqlite3": "^7.6.13",
     "@types/shimmer": "^1.2.0",
     "axios": "^1.13.2",
-    "better-sqlite3": "12.2.0",
     "csv-parse": "5.6.0",
     "date-fns": "4.1.0",
     "dotenv": "^16.6.1",
