+++ src/pages/bookInfo.tsx
@@ -1,6 +1,7 @@
 import { formatDistanceToNow } from "date-fns";
 import { type FC, type PropsWithChildren, Fragment } from "hono/jsx";
 import { useRequestContext } from "hono/jsx-renderer";
+import { endTime, startTime } from "hono/timing";
 import { sql } from "kysely";
 import { BOOK_STATUS, BOOK_STATUS_MAP } from "../constants";
 import type { HiveBook, UserBook } from "../types";
@@ -16,6 +17,7 @@ async function Recommendations({
   did: string | null;
 }) {
   const c = useRequestContext();
+  startTime(c, "db_peer_books");
   const peerBooks = await c
     .get("ctx")
     .db.selectFrom("user_book")
@@ -24,10 +26,13 @@ async function Recommendations({
     .orderBy("indexedAt", "desc")
     .limit(100)
     .execute();
+  endTime(c, "db_peer_books");
 
+  startTime(c, "resolver_dids_to_handles");
   const didHandleMap = await c
     .get("ctx")
     .resolver.resolveDidsToHandles(peerBooks.map((s) => s.userDid));
+  endTime(c, "resolver_dids_to_handles");
 
   if (!peerBooks.length) {
     return (
@@ -287,8 +292,11 @@ export const BookInfo: FC<{
   book: HiveBook;
 }> = async ({ book }) => {
   const c = useRequestContext();
+  startTime(c, "get_session");
   const did = (await c.get("ctx").getSessionAgent())?.did ?? null;
+  endTime(c, "get_session");
 
+  startTime(c, "db_user_book");
   const rawUserBook = did
     ? await c
         .get("ctx")
@@ -298,8 +306,10 @@ export const BookInfo: FC<{
         .where("hiveId", "==", book.id)
         .executeTakeFirst()
     : undefined;
+  endTime(c, "db_user_book");
   const usersBook = rawUserBook ? hydrateUserBook(rawUserBook) : undefined;
 
+  startTime(c, "db_reviews_of_this_book");
   const reviewsOfThisBook = await c
     .get("ctx")
     .db.selectFrom("user_book")
@@ -323,6 +333,7 @@ export const BookInfo: FC<{
     .orderBy("user_book.createdAt", "desc")
     .limit(10_000)
     .execute();
+  endTime(c, "db_reviews_of_this_book");
 
   return (
     <div className="mx-auto max-w-7xl px-4 py-8">
