+++ src/pages/genres.tsx
@@ -27,24 +27,12 @@ export const GenresDirectory: FC = async () => {
   const c = useRequestContext();
 
   startTime(c, "genres-query");
-  // Query all unique genres with their counts
   const genres: GenreWithCount[] = (
-    await (
-      c
-        .get("ctx")
-        .db.selectFrom("hive_book")
-        .select([
-          sql<string>`value`.as("genre"),
-          sql<number>`COUNT(*)`.as("count"),
-        ])
-        .innerJoin(
-          sql`json_each(hive_book.genres)` as any,
-          sql`1` as any,
-          sql`1` as any,
-        ) as any
-    )
-      .where("hive_book.genres", "is not", null)
-      .groupBy(sql`value`)
+    await c
+      .get("ctx")
+      .db.selectFrom("hive_book_genre")
+      .select(["genre", sql<number>`COUNT(*)`.as("count")])
+      .groupBy("genre")
       .orderBy(sql`COUNT(*)`, "desc")
       .execute()
   ).filter((g: GenreWithCount) => g.count > 10);
