+++ src/routes/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Route composition: re-exports main app router and shared search for xrpc/ingester.
+ * Domain routes: pages, profile, books, comments, api; main.tsx composes them.
+ */
+export { searchBooks } from "./lib";
+export { mainRouter, createRouter } from "./main";
